refactor(ui-testing): move ComposeResources to a resource package
